LPS-665 Source formatting
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@23082 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/lar/PortletDataContextImpl.java
@@ -490,9 +490,8 @@ public class PortletDataContextImpl implements PortletDataContext {
 			MBMessage newMessage =
 				MBMessageLocalServiceUtil.addDiscussionMessage(
 					userId, message.getUserName(), classObj.getName(),
-					((Long)newClassPK).longValue(), threadId,
-					parentMessageId, message.getSubject(), message.getBody(),
-					serviceContext);
+					newClassPK, threadId, parentMessageId, message.getSubject(),
+					message.getBody(), serviceContext);
 
 			messagePKs.put(message.getMessageId(), newMessage.getMessageId());
 			threadPKs.put(message.getThreadId(), newMessage.getThreadId());
